requirements.txt
requests==2.31.0
confluent-kafka==2.3.0
python-dotenv==1.0.1
